requirements.txt
Django==2.1.5
django-rest-knox==4.1.0
django-cors-headers==2.2.0
djangorestframework==3.8.2

